# docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaymentTransaction": {
        "example": {
          "amount": 150.0,
          "asset_code": "USDC",
          "confirmed_at": "2026-01-01T01:00:00Z",
          "created_at": "2026-01-01T00:00:00Z",
          "from_address": "SYSTEM_POOL",
          "id": "payment-001",
          "outage_id": "outage-001",
          "retry_count": 0,
          "sla_result_id": 1,
          "status": "confirmed",
          "to_address": "OUTAGE_SETTLEMENT",
          "transaction_hash": "tx-123abc",
          "type": "reward"
        },
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "asset_code": {
            "title": "Asset Code",
            "type": "string"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "from_address": {
            "title": "From Address",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_retried_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Retried At"
          },
          "outage_id": {
            "title": "Outage Id",
            "type": "string"
          },
          "retry_count": {
            "default": 0,
            "title": "Retry Count",
            "type": "integer"
          },
          "sla_result_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Result Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "to_address": {
            "title": "To Address",
            "type": "string"
          },
          "transaction_hash": {
            "title": "Transaction Hash",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "transaction_hash",
          "type",
          "amount",
          "asset_code",
          "from_address",
          "to_address",
          "status",
          "outage_id",
          "created_at"
        ],
        "title": "PaymentTransaction",
        "type": "object"
      },
...
      "ProviderCallbackRequest": {
        "properties": {
          "nonce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonce"
          },
          "provider_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Ref"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "transaction_id",
          "status"
        ],
        "title": "ProviderCallbackRequest",
        "type": "object"
      },
      "ReconcileRequest": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ReconcileRequest",
        "type": "object"
      },
      "ReconciliationHistoryEntry": {
        "description": "A single reconciliation history entry.",
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "new_status": {
            "title": "New Status",
            "type": "string"
          },
          "previous_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Status"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "event_type",
          "new_status",
          "timestamp"
        ],
        "title": "ReconciliationHistoryEntry",
        "type": "object"
      },
      "ReconciliationHistoryResponse": {
        "description": "Payment reconciliation history response.",
        "properties": {
          "current_status": {
            "title": "Current Status",
            "type": "string"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ReconciliationHistoryEntry"
            },
            "title": "History",
            "type": "array"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "transaction_id",
          "current_status",
          "history"
        ],
        "title": "ReconciliationHistoryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "ApexChainx Backend API",
    "title": "ApexChainx API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/payments/": {
      "get": {
        "description": "List payments with filtering and pagination.\n\nSupports both offset-based (page/page_size) and cursor-based (cursor/limit) pagination.\nWhen ``cursor`` is provided, cursor-based pagination is used; otherwise offset-based.",
        "operationId": "list_payments_api_v1_payments__get",
        "parameters": [
          {
            "description": "Page number (offset pagination). Not used when cursor is provided.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (offset pagination). Not used when cursor is provided.",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page.",
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Cursor for cursor-based pagination. Overrides page/page_size.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for cursor-based pagination. Overrides page/page_size.",
              "title": "Cursor"
            }
          },
          {
            "description": "Limit for cursor-based pagination (used with cursor).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Limit for cursor-based pagination (used with cursor).",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "outage_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outage Id"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Payments",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/api/v1/payments/provider-callback": {
      "post": {
        "description": "Inbound callback from a payment provider to update payment status (BE-028).\n\nSecurity model:\n- HMAC-SHA256 signature (X-Webhook-Signature) verified when\n  PAYMENT_WEBHOOK_SECRET is configured.  The signed message includes the\n  nonce so replaying a captured request fails if the nonce changes.\n- Replay protection: the nonce (from X-Callback-Nonce header or the\n  ``nonce`` body field) must be unique within a 5-minute window.\n  Duplicate nonces are rejected with 409 Conflict.\n- Idempotency: a callback that moves a payment into its current status\n  is silently accepted (returns 200 with the unchanged record).\n- Failures and suspicious events are written to the audit log so they\n  are reviewable later.",
        "operationId": "provider_callback_api_v1_payments_provider_callback_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-webhook-signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Signature"
            }
          },
          {
            "in": "header",
            "name": "x-callback-nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Callback-Nonce"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentTransaction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider Callback",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/{transaction_id}": {
      "get": {
        "operationId": "get_payment_api_v1_payments__transaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentTransaction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Payment",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/{transaction_id}/history": {
      "get": {
        "operationId": "get_payment_history_api_v1_payments__transaction_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Payment History Api V1 Payments  Transaction Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Payment History",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/{transaction_id}/reconcile": {
      "post": {
        "operationId": "reconcile_payment_api_v1_payments__transaction_id__reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentTransaction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile Payment",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/{transaction_id}/reconciliation-history": {
      "get": {
        "description": "Get detailed reconciliation history for a payment with timestamps and actor context.\n\nBE-027: Returns a stable, structured history suitable for frontend drawer or audit screen.",
        "operationId": "get_payment_reconciliation_history_api_v1_payments__transaction_id__reconciliation_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Payment Reconciliation History",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/{transaction_id}/retry": {
      "post": {
        "operationId": "retry_payment_api_v1_payments__transaction_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentTransaction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Payment",
        "tags": [
          "payments"
        ]
      }
    },
...
  }
}